refactor: improve modal static methods type definition (#5090)
* refactor: improve modal static methods type definition

* chore: add exports.

* chore: remove unused import
