+++ components/modal/index.tsx
@@ -1,20 +1,9 @@
-import React from 'react';
-import Modal from './Modal';
+import Modal, { ModalFuncProps } from './Modal';
 import confirm from './confirm';
 import assign from 'object-assign';
 
-export interface ModalFuncProps {
-  visible?: boolean;
-  title?: React.ReactNode | string;
-  content?: React.ReactNode | string;
-  onOk?: (func: Function) => any;
-  onCancel?: (func: Function) => any;
-  width?: string | number;
-  iconClassName?: string;
-  okText?: string;
-  cancelText?: string;
-  iconType?: string;
-}
+export { ModalFuncProps }
+
 Modal.info = function (props: ModalFuncProps) {
   const config = assign({}, {
     type: 'info',
